requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
boto3==1.39.17
botocore==1.39.17
python-dotenv==1.1.1
pydantic==2.11.7

